mk-file-protocol-checker: ignore files with no protocol (#11)
* fix a typo and two warnings

* mk-file-protocol-checker: ignore files with no protocol

Right now the predicate *always* returns true if given a filename with
no protocol, even if the protocols set it was created doesn’t include an
empty string.

This means the HTTP(S) file type checker accept any filename with no
protocol. In practice we don’t see the issue because the `:base` checker
is ran before the `:http` one.

This commit extracts the “filename has no protocol” check off the
`mk-protocol-checker` function. The only breaking change is that the
`:base` checker no longer support filenames starting with `://` (with no
protocol before) but I think that’s ok.